test(e2e): single-source the offline stubs and type-check the e2e suite
dismiss-gesture duplicated the thumbnail stub and skipped the font stub,
so the suite's offline guarantee held for one spec file only. helpers.ts
now owns stubGoogleFonts/stubCartoBasemap/stubLocalThumbnails, and
apps/web/e2e joins tsconfig.tools.json.
